Move some tests to new file mocpy/tests/test_interval_set.py
